+++ src/vs/platform/files/node/diskFileSystemProvider.ts
@@ -21,9 +21,9 @@ import { IDirent, Promises, RimRafMode, SymlinkSupport } from 'vs/base/node/pfs'
 import { localize } from 'vs/nls';
 import { createFileSystemProviderError, FileAtomicReadOptions, FileDeleteOptions, FileOpenOptions, FileOverwriteOptions, FileReadStreamOptions, FileSystemProviderCapabilities, FileSystemProviderError, FileSystemProviderErrorCode, FileType, FileWriteOptions, IFileSystemProviderWithFileAtomicReadCapability, IFileSystemProviderWithFileFolderCopyCapability, IFileSystemProviderWithFileReadStreamCapability, IFileSystemProviderWithFileReadWriteCapability, IFileSystemProviderWithOpenReadWriteCloseCapability, isFileOpenForWriteOptions, IStat } from 'vs/platform/files/common/files';
 import { readFileIntoStream } from 'vs/platform/files/common/io';
-import { AbstractNonRecursiveWatcherClient, AbstractUniversalWatcherClient, IDiskFileChange, ILogMessage, isRecursiveWatchRequest, IUniversalWatcheRequest } from 'vs/platform/files/common/watcher';
+import { AbstractNonRecursiveWatcherClient, AbstractUniversalWatcherClient, IDiskFileChange, ILogMessage } from 'vs/platform/files/common/watcher';
 import { ILogService } from 'vs/platform/log/common/log';
-import { AbstractDiskFileSystemProvider } from 'vs/platform/files/common/diskFileSystemProvider';
+import { AbstractDiskFileSystemProvider, IDiskFileSystemProviderOptions } from 'vs/platform/files/common/diskFileSystemProvider';
 import { toErrorMessage } from 'vs/base/common/errorMessage';
 import { UniversalWatcherClient } from 'vs/platform/files/node/watcher/watcherClient';
 import { NodeJSWatcherClient } from 'vs/platform/files/node/watcher/nodejs/nodejsClient';
@@ -40,31 +40,6 @@ import { NodeJSWatcherClient } from 'vs/platform/files/node/watcher/nodejs/nodej
 	}
 })();
 
-export interface IWatcherOptions {
-
-	/**
-	 * If `true`, will enable polling for all watchers, otherwise
-	 * will enable it for paths included in the string array.
-	 *
-	 * @deprecated this only exists for WSL1 support and should never
-	 * be used in any other case.
-	 */
-	usePolling: boolean | string[];
-
-	/**
-	 * If polling is enabled (via `usePolling`), defines the duration
-	 * in which the watcher will poll for changes.
-	 *
-	 * @deprecated this only exists for WSL1 support and should never
-	 * be used in any other case.
-	 */
-	pollingInterval?: number;
-}
-
-export interface IDiskFileSystemProviderOptions {
-	watcher?: IWatcherOptions;
-}
-
 export class DiskFileSystemProvider extends AbstractDiskFileSystemProvider implements
 	IFileSystemProviderWithFileReadWriteCapability,
 	IFileSystemProviderWithOpenReadWriteCloseCapability,
@@ -74,9 +49,9 @@ export class DiskFileSystemProvider extends AbstractDiskFileSystemProvider imple
 
 	constructor(
 		logService: ILogService,
-		private readonly options?: IDiskFileSystemProviderOptions
+		options?: IDiskFileSystemProviderOptions
 	) {
-		super(logService);
+		super(logService, options);
 	}
 
 	//#region File Capabilities
@@ -615,25 +590,6 @@ export class DiskFileSystemProvider extends AbstractDiskFileSystemProvider imple
 		verboseLogging: boolean
 	): AbstractUniversalWatcherClient {
 		return new UniversalWatcherClient(changes => onChange(changes), msg => onLogMessage(msg), verboseLogging);
-	}
-
-	protected override massageWatchRequests(requests: IUniversalWatcheRequest[]): void {
-		const usePolling = this.options?.watcher?.usePolling;
-		if (usePolling === true) {
-			for (const request of requests) {
-				if (isRecursiveWatchRequest(request)) {
-					request.pollingInterval = this.options?.watcher?.pollingInterval ?? 5000;
-				}
-			}
-		} else if (Array.isArray(usePolling)) {
-			for (const request of requests) {
-				if (isRecursiveWatchRequest(request)) {
-					if (usePolling.includes(request.path)) {
-						request.pollingInterval = this.options?.watcher?.pollingInterval ?? 5000;
-					}
-				}
-			}
-		}
 	}
 
 	protected createNonRecursiveWatcher(
